Consolidate the test scaffolding
Move the Process handling capabilities and the Provisioning of test
artifacts (test files, directories, environment variables) into
a single ecflow/test/scaffold support library.
