+++ libs/server/test/TestServerConfigurations.cpp
@@ -32,6 +32,8 @@ private:
 
 BOOST_AUTO_TEST_SUITE(U_Server)
 
+using namespace ecf::test::scaffold;
+
 /*
  * The following exports an environment variable used for tests, which changes the location of the SSL certificates.
  * Instead of the default location (HOME/.ecflowrc/ssl), we use the current test directory.
